feat: 정책 추천 순위 계산 컴포넌트 분리 (#56)
