+++ src/main/java/com/chungbazi/server/domain/policy/application/PersonalizedPolicyService.java
@@ -1,7 +1,7 @@
 package com.chungbazi.server.domain.policy.application;
 
 import com.chungbazi.server.domain.policy.application.dto.PolicyRecommendationContext;
-import com.chungbazi.server.domain.policy.application.support.PersonalizedPolicyScorer;
+import com.chungbazi.server.domain.policy.application.support.PersonalizedPolicyRanker;
 import com.chungbazi.server.domain.policy.domain.entity.Policy;
 import com.chungbazi.server.domain.policy.domain.repository.PolicyLikeRepository;
 import com.chungbazi.server.domain.policy.domain.repository.RecentViewedPolicyRepository;
@@ -16,11 +16,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.EnumMap;
 import java.util.List;
-import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
@@ -29,13 +25,12 @@ public class PersonalizedPolicyService {
 
     private static final int CANDIDATE_SIZE = 300;
     private static final int BEHAVIOR_HISTORY_SIZE = 50;
-    private static final int MAX_SAME_CATEGORY_COUNT = 3;
 
     private final PolicyRepository policyRepository;
     private final UserInterestRepository userInterestRepository;
     private final PolicyLikeRepository policyLikeRepository;
     private final RecentViewedPolicyRepository recentViewedPolicyRepository;
-    private final PersonalizedPolicyScorer scorer;
+    private final PersonalizedPolicyRanker personalizedPolicyRanker;
 
     public List<Policy> getPersonalizedPolicyEntities(User user, int size) {
         // TODO: 추후 캐싱 고려
@@ -62,21 +57,7 @@ public class PersonalizedPolicyService {
                 )
         );
 
-        List<Policy> scoredPolicies = candidates.stream()
-                .filter(policy -> scorer.isEligible(user, policy))
-                .sorted(Comparator
-                        .comparingInt((Policy policy) -> scorer.score(user, context, policy))
-                        .reversed()
-                        .thenComparing(Policy::getRegisteredAt, Comparator.reverseOrder()))
-                .toList();
-
-        // 관심 대분류를 하나만 선택한 경우, 카테고리 다양성 제한 없이 점수순 그대로
-        if (context.interestCategoryCounts().size() <= 1) {
-            return scoredPolicies.stream()
-                    .limit(size)
-                    .toList();
-        }
-        return diversifyByCategory(scoredPolicies, size);
+        return personalizedPolicyRanker.rank(user, context, candidates, size);
     }
 
     public List<Policy> getPersonalizedPolicyEntities(User user, PolicyCategoryType category, int size) {
@@ -112,47 +93,6 @@ public class PersonalizedPolicyService {
                 )
         );
 
-        return candidates.stream()
-                .filter(policy -> scorer.isEligible(user, policy))
-                .sorted(Comparator
-                        .comparingInt((Policy policy) -> scorer.score(user, context, policy))
-                        .reversed()
-                        .thenComparing(Policy::getRegisteredAt, Comparator.reverseOrder()))
-                .limit(size)
-                .toList();
-    }
-
-    private List<Policy> diversifyByCategory(List<Policy> policies, int size) {
-        List<Policy> selectedPolicies = new ArrayList<>();
-        Map<PolicyCategoryType, Integer> selectedCategoryCounts = new EnumMap<>(PolicyCategoryType.class);
-
-        // 관심 대분류가 여러 개인 경우, 같은 대분류가 과도하게 몰리지 않도록 노출 개수 제한
-        for (Policy policy : policies) {
-            PolicyCategoryType category = policy.getCategory();
-            int categoryCount = selectedCategoryCounts.getOrDefault(category, 0);
-
-            if (categoryCount >= MAX_SAME_CATEGORY_COUNT) {
-                continue;
-            }
-            selectedPolicies.add(policy);
-            selectedCategoryCounts.put(category, categoryCount + 1);
-
-            if (selectedPolicies.size() == size) {
-                return selectedPolicies;
-            }
-        }
-
-        // 제한 때문에 목표 개수를 채우지 못한 경우에는 점수순으로 남은 정책 추가
-        for (Policy policy : policies) {
-            if (selectedPolicies.contains(policy)) {
-                continue;
-            }
-            selectedPolicies.add(policy);
-
-            if (selectedPolicies.size() == size) {
-                return selectedPolicies;
-            }
-        }
-        return selectedPolicies;
+        return personalizedPolicyRanker.rank(user, context, candidates, size);
     }
 }
