+++ src/test/java/com/chungbazi/server/domain/policy/application/support/PersonalizedPolicyScorerTest.java
@@ -24,7 +24,7 @@ public class PersonalizedPolicyScorerTest {
 
     @BeforeEach
     void setUp() {
-        scorer = new PersonalizedPolicyScorer();
+        scorer = new PersonalizedPolicyScorer(new PolicyIncomeMatcher());
     }
 
     @Test
